chore:move icons to assets
