+++ frontend/src/assets/icons/AgentsIcon.tsx
