+++ frontend/src/assets/icons/DashboardIcon.tsx
