+++ frontend/src/assets/icons/EmptyIcon.tsx
