+++ frontend/src/assets/icons/Logo.tsx
