+++ frontend/src/assets/icons/LogsIcon.tsx
