+++ frontend/src/assets/icons/MyAgentIcon.tsx
