+++ frontend/src/assets/icons/Polygon.tsx
