+++ frontend/src/assets/icons/SearchIcon.tsx
