+++ frontend/src/assets/icons/TemplateEditIcon.tsx
