+++ frontend/src/assets/icons/TemplatesIcon.tsx
