+++ frontend/src/assets/icons/UsersIcon.tsx
