+++ frontend/src/components/atoms/DropDownMenu.tsx
@@ -5,7 +5,7 @@ import * as React from 'react';
 import * as DropdownMenuPrimitive from '@radix-ui/react-dropdown-menu';
 import { Check, ChevronRight, Circle } from 'lucide-react';
 
-import PolygonIcon from '../icons/Polygon';
+import PolygonIcon from '@app/assets/icons/Polygon';
 import { cn } from '../lib/utils';
 
 const DropdownMenu = DropdownMenuPrimitive.Root;
